fix(backup/destination/fs): skip directory fsync on Windows
Windows returns ERROR_ACCESS_DENIED when Sync is called on a directory
handle. NTFS/ReFS rename is journaled, so the Unix-style "fsync parent
after rename for durability" ceremony is neither needed nor supported.
Split fsyncDir into build-tagged variants.
